pyproject.toml
[tool.poetry]
name = "law-chatbot-poc"
version = "0.1.0"
description = ""
authors = ["KalenHsu <KalenHsu@going.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "1.25.0"
langchain = "0.0.244"
chromadb = "0.4.3"
openai = "0.27.8"
html2text = "2020.1.16"
google-api-core = "2.11.1"
google-api-python-client = "2.95.0"
google-auth = "2.22.0"
google-auth-httplib2 = "0.1.0"
googleapis-common-protos = "1.59.1"
tiktoken = "0.4.0"
faiss-cpu = "1.7.4"
python-dotenv = "1.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
